changed (unsigned) to (NSInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11743 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/ORCARootService/ORCARootService.h
@@ -92,7 +92,7 @@
 - (void) cancelRequest:(NSNotification*)aNote;
 - (void) requestNotification:(NSNotification*)aNote;
 - (void) sendRequest:(NSMutableDictionary*)request fromObject:(id)anObject;
-- (unsigned) connectionHistoryCount;
+- (NSUInteger) connectionHistoryCount;
 - (id) connectionHistoryItem:(NSUInteger)index;
 - (void) clearFitInFlight;
